Feat/11 add product option

+++ src/main/java/com/kfreemarket/reemarket_server/domain/product/dto/ProductDto.java
@@ -13,7 +13,7 @@ public class ProductDto {
 
     private String productName;
 
-    private Integer productPrice;
+    private Long productPrice;
 
     private Integer stock;
